fix: modify API 'remove the orderline' based on isStockable

+++ convex/orderlines.ts
@@ -1,13 +1,9 @@
 import { getAuthUserId } from "@convex-dev/auth/server"
 import { ConvexError, v } from "convex/values"
 import { upsertOrderlineSchema } from "../types/schema/orderline-schema"
+import { Id } from "./_generated/dataModel"
 import { mutation, query } from "./_generated/server"
-import {
-  cashierProcedure,
-  managerProcedure,
-  protectedProcedure,
-  zMutation,
-} from "./helpers"
+import { managerProcedure, protectedProcedure, zMutation } from "./helpers"
 
 export const findAll = query({
   args: {
@@ -531,16 +527,34 @@ export const remove = mutation({
     countInStock: v.number(),
   },
   handler: async (ctx, { id, productId, countInStock }) => {
-    await cashierProcedure(ctx)
+    // cashierProcedure(ctx, {})
+    const userId = await getAuthUserId(ctx)
+    const user = userId !== null ? await ctx.db.get(userId) : null
+    if (
+      user?.role !== "DEWA" &&
+      user?.role !== "ADMIN" &&
+      user?.role !== "CASHIER"
+    )
+      throw new ConvexError("You do not have access!")
 
+    const company = await ctx.db.get(user.companyId as Id<"companies">)
+    if (!company) {
+      throw new ConvexError("Company not found!")
+    }
+
+    // const product = await ctx.db.get(productId)
     const removeOrderline = await ctx.db.delete(id)
-    const updateStock = await ctx.db.patch(productId, {
-      countInStock,
-    })
+
+    let updateStock = null
+    if (company.isStockable) {
+      updateStock = await ctx.db.patch(productId, {
+        countInStock,
+      })
+    }
 
     return {
       removeOrderline,
-      updateStock,
+      ...(updateStock ? { updateStock } : {}),
     }
   },
 })
